src/scraper.py
"""Scraper to fetch app data from Play Store or mock data."""

import requests
from bs4 import BeautifulSoup

class AppScraper:
    def __init__(self):
        self.base_url = "https://play.google.com/store/search"
    
    def search_apps(self, brand_name, mock=True):
        """
        Search for apps by brand name.
        
        Args:
            brand_name: Brand to search for
            mock: If True, returns mock data for demo
        
        Returns:
            List of app dictionaries
        """
        if mock:
            return self._get_mock_data(brand_name)
        
        # Real scraping implementation (optional)
        return self._scrape_play_store(brand_name)
    
    def _get_mock_data(self, brand_name):
        """Generate mock app data for demonstration."""
        mock_apps = [
            {
                "name": "PhonePe",
                "package_id": "com.phonepe.app",
                "publisher": "PhonePe Private Limited",
                "downloads": "500M+",
                "rating": 4.5,
                "description": "Official PhonePe app for payments"
            },
            {
                "name": "PhonPe - UPI Payments",
                "package_id": "com.phonepe.app.update",
                "publisher": "PhonePe Services",
                "downloads": "100K+",
                "rating": 4.2,
                "description": "Fast UPI payments and recharges"
            },
            {
                "name": "PhonePe Secure",
                "package_id": "com.phonepe.secure.app",
                "publisher": "Mobile Payment Solutions",
                "downloads": "50K+",
                "rating": 3.8,
                "description": "Secure payment app for PhonePe users"
            }
        ]
        return mock_apps
    
    def _scrape_play_store(self, brand_name):
        """Actual Play Store scraping (implement if needed)."""
        # TODO: Implement real scraping
        # Note: Respect rate limits and ToS
        pass
