requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
Pillow==10.1.0
imagehash==4.3.1
